+++ rust/crates/sift_cli/src/cmd/import/mcap/mod.rs
@@ -1,0 +1,6 @@
+mod detect_mcap_config;
+pub mod import;
+mod ros2_schema;
+
+#[cfg(test)]
+mod tests;