+++ rust/crates/sift_cli/src/main.rs
@@ -143,6 +143,7 @@ fn run(clargs: cli::Args) -> Result<ExitCode> {
                 }
             },
             cli::ImportCmd::Ulog(args) => run_future(cmd::import::ulog::import::run(ctx, args)),
+            cli::ImportCmd::Mcap(args) => run_future(cmd::import::mcap::import::run(ctx, args)),
             cli::ImportCmd::Backup(args) => match args.cmd {
                 Some(cli::BackupCmd::Ls(ls_args)) => {
                     run_future(cmd::import::backup::ls(ctx, ls_args))